+++ src/Aptabase.Core/LocalHttpsClientHandler.cs
@@ -1,0 +1,13 @@
+namespace Aptabase.Core;
+
+public class LocalHttpsClientHandler : HttpClientHandler
+{
+    public LocalHttpsClientHandler()
+    {
+        // Disable SSL verification
+        ServerCertificateCustomValidationCallback = (sender, cert, chain, sslPolicyErrors) =>
+        {
+            return true;
+        };
+    }
+}